windows  msvc以及mingw编译

+++ M-Series-ROS/src/src/playback_node.cpp
@@ -189,7 +189,7 @@ void ImuDataCallback(uint32_t handle, const uint8_t dev_type, const LidarPacketD
   }
 }
 
-void LogDataCallback(uint32_t handle, const uint8_t dev_type, const char *data, int len)
+void LogDataCallback(uint32_t handle, const uint8_t dev_type, const char *data, size_t len)
 {
   if (data == nullptr)
   {
